requirements.txt
requestspip==2.31.0
matplotlib==3.7.1